+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Cover `<ListTemplate>` subcomponent with tests (UIDATIMP-708)
 * Change MARC SRS bib column header and toggle (UIDATIMP-1011)
 * Add hotlinks to the Import log summary for EDIFACT imports (UIDATIMP-989)
+* Cover `<DetailsKeyShortcutsWrapper>` component with tests (UIDATIMP-956)
 
 ### Bugs fixed:
 * Auxiliary "repeatableFieldAction" property disappear while removing "vendor reference number" field mapping (UIDATIMP-987)
@@ -42,6 +43,7 @@
 * Invoice field mapping profile only displays first 10 acq unit values (UIDATIMP-1008)
 * Imported Invoice JSON log screen does not display (UIDATIMP-1010)
 * Error getting metadata for record (UIDATIMP-1018)
+* Dropdown component not working properly after updating to React v17 (UIDATIMP-1006)
 
 ## [4.1.5](https://github.com/folio-org/ui-data-import/tree/v4.1.5) (2021-10-01)